Require PASS property for shouldGetLastChanges test
Assembla now requires authenticated access for their Subversion
repositories.  The shouldGetLastChangesFromLatestTag test already
used this same technique to only run the test when the PASS system
property is provided.  Make shouldGetLastChanges the same as
shouldGetLastChangesFromLatestTag.

Test fails on the master branch as well.

This change ignores shouldGetLastChanges on ci.jenkins.io in the same
way that shouldGetLastChangesFromLatestTag was already ignored.